flask_taller/requirements.txt
alembic==0.8.9
aniso8601==8.1.1
bcrypt==3.1.2
cffi==1.9.1
click==6.6
coverage==4.2
Flask==0.12
Flask-Bcrypt==0.7.1
flask-marshmallow==0.14.0
Flask-Migrate==2.0.2
Flask-RESTful==0.3.8
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
pkg-resources==0.0.0
psycopg2==2.8.6
pycparser==2.17
PyJWT==1.4.2
python-editor==1.0.3
pytz==2021.1
six==1.10.0
SQLAlchemy==1.3.23
Werkzeug==0.11.13
